Create role-specific permissions for products

+++ Controllers/ProductsController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using OnlineShopDAW.Models;
+using Microsoft.AspNet.Identity;
 
 namespace OnlineShopDAW.Controllers
 {
@@ -16,6 +17,9 @@ namespace OnlineShopDAW.Controllers
             var products = from prod
                            in db.Products.Include("Category").Include("Reviews")
                            select prod;
+
+            ViewBag.esteAdmin = User.IsInRole("administrator");
+            ViewBag.esteColaborator = User.IsInRole("collaborator");
             ViewBag.Products = products;
             return View();
         }
@@ -26,10 +30,13 @@ namespace OnlineShopDAW.Controllers
                 .Include("Category")
                 .Include("Reviews")
                 .FirstOrDefault(p => p.ProductId == id);
+
+            SetButtonVisibility(product);
             ViewBag.Product = product;
             return View(product);
         }
 
+        [Authorize(Roles = "administrator,collaborator")]
         public ActionResult New()
         {
             Product product = new Product();
@@ -39,15 +46,26 @@ namespace OnlineShopDAW.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = "administrator,collaborator")]
         public ActionResult New(Product product)
         {
             try
             {
                 product.CreatedAt = DateTime.UtcNow;
+                product.UserId = User.Identity.GetUserId();
                 product.Categ = GetAllCategories();
+                product.Status = ProductStatus.pending;
+
+                TempData["message"] = "Produsul a fost trimis catre evaluare! Va multumim!";
+
+                if (User.IsInRole("administrator"))
+                {
+                    TempData["message"] = "Produsul a fost adaugat!";
+                    product.Status = ProductStatus.accepted;
+                }
+
                 db.Products.Add(product);
                 db.SaveChanges();
-                TempData["message"] = "Produsul a fost trimis catre evaluare! Va multumim!";
                 return RedirectToAction("Index");
             }
             catch (Exception e)
@@ -57,6 +75,7 @@ namespace OnlineShopDAW.Controllers
             }
         }
 
+        [Authorize(Roles = "administrator,collaborator")]
         public ActionResult Edit(int id)
         {
             var product = db.Products
@@ -64,29 +83,49 @@ namespace OnlineShopDAW.Controllers
                 .Include("Reviews")
                 .FirstOrDefault(p => p.ProductId == id);
             product.Categ = GetAllCategories();
-            ViewBag.Product = product;
-            return View(product);
+            if (product.UserId == User.Identity.GetUserId() || User.IsInRole("administrator"))
+            {
+                ViewBag.Product = product;
+                return View(product);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa modificati produse!";
+                return Redirect("/products");
+            }
         }
 
         [HttpPut]
+        [Authorize(Roles = "administrator,collaborator")]
         public ActionResult Edit(int id, Product requestProduct)
         {
             try
             {
                 Product product = db.Products.Find(id);
-                if (TryUpdateModel(product))
+
+                if (product.UserId == User.Identity.GetUserId() || User.IsInRole("administrator"))
                 {
-                    product.Name = requestProduct.Name;
-                    product.Description = requestProduct.Description;
-                    product.Price = requestProduct.Price;
-                    product.DiscountedPrice = requestProduct.DiscountedPrice;
-                    product.Image = requestProduct.Image;
-                    product.Stock = requestProduct.Stock;
-                    product.Status = requestProduct.Status;
-                    product.Category = requestProduct.Category;
-                    db.SaveChanges();
+                    if (TryUpdateModel(product))
+                    {
+                        product.Name = requestProduct.Name;
+                        product.Description = requestProduct.Description;
+                        product.Price = requestProduct.Price;
+                        product.DiscountedPrice = requestProduct.DiscountedPrice;
+                        product.Image = requestProduct.Image;
+                        product.Stock = requestProduct.Stock;
+                        product.Status = requestProduct.Status;
+                        product.Category = requestProduct.Category;
+                        product.Status = requestProduct.Status;
+                        db.SaveChanges();
+                        TempData["message"] = "Produsul a fost modificat!";
+                    }
+                    return Redirect("/products/show/" + product.ProductId);
                 }
-                return RedirectToAction("Index");
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa modificati acest produs!";
+                    return Redirect("/products");
+                }
             }
             catch (Exception e)
             {
@@ -96,12 +135,23 @@ namespace OnlineShopDAW.Controllers
         }
 
         [HttpDelete]
+        [Authorize(Roles = "administrator,collaborator")]
         public ActionResult Delete(int id)
         {
-            Product product = db.Products.Find(id);
-            db.Products.Remove(product);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+                Product product = db.Products.Find(id);
+            if (product.UserId == User.Identity.GetUserId() || User.IsInRole("administrator"))
+            {
+
+                db.Products.Remove(product);
+                db.SaveChanges();
+                TempData["message"] = "Produsul a fost sters!";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti acest produs!";
+                return RedirectToAction("Index");
+            }
         }
 
         [NonAction]
@@ -123,6 +173,19 @@ namespace OnlineShopDAW.Controllers
             }
 
             return selectList;
+        }
+
+        private void SetButtonVisibility(Product product)
+        {
+            ViewBag.afisareButoane = false;
+            if (product.UserId == User.Identity.GetUserId() || User.IsInRole("administrator"))
+            {
+                ViewBag.afisareButoane = true;
+            }
+
+            ViewBag.esteAdmin = User.IsInRole("administrator");
+            ViewBag.esteLogat = (User.IsInRole("administrator") || User.IsInRole("collaborator") || User.IsInRole("registered"));
+            ViewBag.utilizatorCurent = User.Identity.GetUserId();
         }
     }
 }
